고정 화면 추가 , main 데이터 생성, shadow 이미지 적용, layout 수정, 간격 수정, 에러 수정
