+++ core/data/src/main/java/com/dnd_9th_3_android/gooding/data/feed/MainTab.kt
@@ -1,0 +1,4 @@
+package com.dnd_9th_3_android.gooding.data.feed
+
+
+val pages = listOf("NOW", "추천")
