+++ core/data/src/main/res/values-hdpi/dimens.xml
@@ -1,0 +1,69 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <!-- top -->
+    <dimen name="top_size">56dp</dimen>
+    <dimen name="top_margin">53dp</dimen>
+    <dimen name="padding_16">16dp</dimen>
+    <dimen name="padding_17">17dp</dimen>
+    <dimen name="padding_24">24dp</dimen>
+    <dimen name="padding_21">21dp</dimen>
+    <dimen name="main_text_sp">18sp</dimen>
+    <dimen name="icon_size">30dp</dimen>
+    <!--level-->
+    <dimen name="text_12">12sp</dimen>
+    <dimen name="border_size">1dp</dimen>
+    <dimen name="padding_8">8dp</dimen>
+    <dimen name="padding_10">10dp</dimen>
+    <dimen name="padding_6">6dp</dimen>
+    <dimen name="corner_4">4dp</dimen>
+    <dimen name="corner_6">6dp</dimen>
+    <dimen name="corner_8">8dp</dimen>
+    <dimen name="level_image_w">166dp</dimen>
+    <dimen name="level_image_h">147dp</dimen>
+    <!--profile-->
+    <dimen name="top_margin_profile">30dp</dimen>
+    <dimen name="padding_20">20dp</dimen>
+    <dimen name="padding_23">23dp</dimen>
+    <dimen name="padding_28">28dp</dimen>
+    <dimen name="profile_image_size">36dp</dimen>
+    <dimen name="text_16_sp">16sp</dimen>
+    <!--tab-->
+    <dimen name="tab_75">75dp</dimen>
+    <dimen name="tab_48">48dp</dimen>
+    <dimen name="text_w_1">54dp</dimen>
+    <dimen name="text_h">18dp</dimen>
+    <dimen name="text_w_2">77dp</dimen>
+    <dimen name="border_size_2">2dp</dimen>
+    <dimen name="arrow_size">24dp</dimen>
+    <dimen name="padding_13">13dp</dimen>
+    <dimen name="padding_33">33dp</dimen>
+    <dimen name="padding_34">34dp</dimen>
+    <dimen name="padding_36">36dp</dimen>
+    <dimen name="padding_15">15dp</dimen>
+    <dimen name="padding_18">18dp</dimen>
+    <dimen name="tab_width">350dp</dimen>
+    <dimen name="start_height">200dp</dimen>
+    <dimen name="text_width">80dp</dimen>
+    <!--content-->
+    <dimen name="zero">0dp</dimen>
+    <dimen name="text_14_sp">14sp</dimen>
+    <dimen name="padding_2">2dp</dimen>
+    <dimen name="size_144">144dp</dimen>
+    <dimen name="padding_12">12dp</dimen>
+    <dimen name="padding_80">80dp</dimen>
+    <dimen name="padding_11_5">11.5dp</dimen>
+    <!--feed-->
+    <dimen name="size_26">26dp</dimen>
+    <dimen name="padding_14">14dp</dimen>
+    <dimen name="padding_4">4dp</dimen>
+    <dimen name="padding_11">11dp</dimen>
+    <dimen name="corner_20">20dp</dimen>
+    <dimen name="image_h">333dp</dimen>
+    <dimen name="image_w">218dp</dimen>
+    <dimen name="text_w_content">268dp</dimen>
+    <dimen name="padding_5_5">5.5dp</dimen>
+
+    <!--main feed-->
+    <dimen name="padding_55">55dp</dimen>
+    <dimen name="book_padding_bottom">197dp</dimen>
+</resources>
