+++ core/data/src/main/res/values-xhdpi/dimens.xml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources></resources>
