--- feature/feed/src/main/res/values-hdpi/dimens.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <!-- tab -->
-    <dimen name="selected_icon_size">5dp</dimen>
-    <dimen name="margin_tab_10">18dp</dimen>
-    <dimen name="margin_tab">18dp</dimen>
-    <dimen name="user_image_size">34dp</dimen>
-    <dimen name="top_space">36dp</dimen>
-    <dimen name="tab_height">56dp</dimen>
-    <dimen name="tab_width">150dp</dimen>
-    <dimen name="size_24">24dp</dimen>
-</resources>
