--- feature/feed/src/main/res/values-xhdpi/dimens.xml
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources></resources>
